requirements.txt
git+https://github.com/potatolondon/djangae.git#egg=djangae
django>=1.8,<1.9
django-secure
django-csp
git+https://github.com/adamalton/django-csp-reports.git#egg=cspreports
django-session-csrf
nose
six
django-extensions
crispy-forms-foundation
pytest
pytest-pep8
factory-boy